less memory with long

+++ src/model/RolloutStorage.py
@@ -8,9 +8,9 @@ class RolloutStorage(object):
         self.num_steps = num_steps
         self.state_shape = state_shape
 
-        self.states = torch.zeros(num_steps + 1, *state_shape)
-        self.rewards = torch.zeros(num_steps, num_agents)
-        self.masks = torch.ones(num_steps + 1, num_agents)
+        self.states = torch.zeros(num_steps + 1, *state_shape).long()
+        self.rewards = torch.zeros(num_steps, num_agents).long()
+        self.masks = torch.ones(num_steps + 1, num_agents).long()
         self.actions = torch.zeros(num_steps, num_agents).long()
         self.values = torch.zeros(num_steps + 1, num_agents).long()
         self.returns = torch.zeros(num_steps + 1, num_agents)
